# docs/swagger.yaml
basePath: /
definitions:
  api.AddMembersRequest:
    properties:
      usernames:
        example:
        - '["bob"'
        - '"charlie"]'
        items:
          type: string
        type: array
    type: object
  api.CreateChannelRequest:
    properties:
      is_private:
        example: false
        type: boolean
      members:
        example:
        - '["user-id-1"'
        - '"user-id-2"]'
        items:
          type: string
        type: array
      name:
        example: Design Team
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
  api.LoginRequest:
    properties:
      password:
        example: s3cur3P@ss!
        type: string
      username:
        example: alice
        type: string
    type: object
  api.LoginResponse:
    properties:
      token:
        example: Bearer eyJhb...
        type: string
      user_id:
        example: a3f4...
        type: string
      username:
        example: alice
        type: string
    type: object
  api.RegisterRequest:
    properties:
      password:
        example: s3cur3P@ss!
        type: string
      username:
        example: alice
        type: string
    type: object
  api.RegisterResponse:
    properties:
      id:
        example: a3f4...
        type: string
      username:
        example: alice
        type: string
    type: object
  api.StatusResponse:
    properties:
      status:
        example: success
        type: string
    type: object
  api.UploadFileResponse:
    properties:
      file_id:
        type: string
      mime:
        type: string
      size:
        type: integer
      uploaded:
        type: string
    type: object
  api.UserResponse:
    properties:
      domain:
        example: localhost
        type: string
      id:
        example: user-1
        type: string
      status:
        example: online
        type: string
      username:
        example: alice
        type: string
    type: object
  models.Channel:
    properties:
      aes_key_enc:
        description: Channel AES-256 key encrypted with server KEK
        items:
          type: integer
        type: array
      created_at:
        type: string
      creator_id:
        type: string
      id:
        type: string
      invite_link:
        description: Unique public/private join link
        type: string
      is_private:
        description: If true, hidden from public search
        type: boolean
      members:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  models.Message:
    properties:
      channel_id:
        type: string
      ciphertext:
        description: AES-256-GCM encrypted payload
        items:
          type: integer
        type: array
      id:
        type: string
      nonce:
        description: GCM nonce (also prepended in ciphertext)
        items:
          type: integer
        type: array
      sender_id:
        type: string
      sent_at:
        type: string
      sig:
        description: 'Reserved: future sender signature'
        items:
          type: integer
        type: array
      type:
        allOf:
        - $ref: '#/definitions/models.MessageType'
        description: '"text" | "file" | "voice"'
    type: object
  models.MessageType:
    enum:
    - text
    - file
    - voice
    type: string
    x-enum-varnames:
    - MessageTypeText
    - MessageTypeFile
    - MessageTypeVoice
  models.Presence:
    properties:
      current_status:
        description: '"online" | "away" | "offline"'
        type: string
      last_seen:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: NodeTalk Support
    url: https://github.com/nodetalk/nodetalk
  description: Self-hosted, end-to-end encrypted communication platform.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://nodetalk.local/terms
  title: NodeTalk API
  version: 0.1.0
paths:
  /api/channels:
    get:
      description: Without ?q=, returns your joined channels. With ?q=, searches across
        all public channels globally.
      parameters:
      - description: Search query for ALL public channels
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Channel'
            type: array
      security:
      - BearerAuth: []
      summary: List channels for the user OR search global public channels
      tags:
      - channels
    post:
      consumes:
      - application/json
      parameters:
      - description: Channel payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Channel'
      security:
      - BearerAuth: []
      summary: Create a channel or direct message
      tags:
      - channels
  /api/channels/{id}:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Channel'
      security:
      - BearerAuth: []
      summary: Get a channel by ID
      tags:
      - channels
  /api/channels/{id}/members:
    get:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.UserResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List members of a channel
      tags:
      - channels
    post:
      consumes:
      - application/json
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Members to add (by username)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AddMembersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      security:
      - BearerAuth: []
      summary: Add multiple members to a channel by username
      tags:
      - channels
  /api/channels/{id}/members/{uid}:
    delete:
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      security:
      - BearerAuth: []
      summary: Remove a member from a channel
      tags:
      - channels
  /api/channels/{id}/messages:
    get:
      description: Returns messages newest-first. Supports cursor pagination.
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Max messages to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Unix timestamp (nanoseconds) cursor for pagination
        in: query
        name: before
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
      security:
      - BearerAuth: []
      summary: List recent messages for a channel
      tags:
      - messages
  /api/files/{id}:
    get:
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
      security:
      - BearerAuth: []
      summary: Download a file by ID
      tags:
      - files
  /api/join/{link}:
    post:
      parameters:
      - description: Unique Channel Invite Link
        in: path
        name: link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      security:
      - BearerAuth: []
      summary: Join a channel via a unique invite link
      tags:
      - channels
  /api/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Log in and obtain a session token
      tags:
      - auth
  /api/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      security:
      - BearerAuth: []
      summary: Log out and invalidate session
      tags:
      - auth
  /api/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - users
  /api/register:
    post:
      consumes:
      - application/json
      description: Creates a new NodeTalk user account.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Register a new account
      tags:
      - auth
  /api/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File to upload (max 50 MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.UploadFileResponse'
      security:
      - BearerAuth: []
      summary: Upload a file or voice note
      tags:
      - files
  /api/users:
    get:
      parameters:
      - description: Search query to filter by username
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.UserResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Search globally for users by username
      tags:
      - users
  /api/users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - users
  /api/users/{id}/presence:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Presence'
      security:
      - BearerAuth: []
      summary: Get presence status for a user
      tags:
      - users
  /api/users/me:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
      security:
      - BearerAuth: []
      summary: Delete current user account
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"